fix(arcgis-rest-auth): ensure that mixed casing of federated server urls does not break the system
When federated servers are configured, the UI does not enforce lower casing of the federated server
name. This can result in hosted services with UPPERCASE urls. This PR ensures that the user session
works with lower-cased urls, and that regardless of casing of inbound requests, the federated token
cache works correctly.

AFFECTS PACKAGES:
@esri/arcgis-rest-auth