remove incorrect lerna instructions
After re-reading the [lerna docs on fixed mode](https://github.com/lerna/lerna#fixedlocked-mode-default), I _think_ lerna is doing the right thing, in particular:

> When you run lerna publish, if a module has been updated since the last time a release was made, it will be updated to the new version you're releasing.

I think the ones that are not getting updated are the ones that have not been updated and do not depend on any packages that have been updated. For example, I'm in the middle of cutting a release now. The only package w/ changes is arcgis-rest-portal. When I run `npm run release:prepare` w/ the minor option, the only packages whose versions are updated from 2.9.0 to 2.10.0 are arcgis-rest-portal and those that depend on it. The "lower" level packages like arcgis-rest-request are untouched. That works b/c the peerDependency range for the ones that did bump is still `^2.0.0`.