feat(UserSession): added optional 'popupWindowFeatures' to 'IOAuth2Options'
Adding an optional 'popupWindowFeatures' property to ‘IOAuth2Options’ allows developers to pass
window features to the popup window. This allows a developer to center the popup if desired using
his own calculated offsets. Currently the popup window features are hard coded. This change simply
exposes them to the developer and provides a default (the prior hard coded values) if not supplied.
The ArcGIS JavaScript API already has the feature:
https://developers.arcgis.com/javascript/latest/api-reference/esri-identity-OAuthInfo.html#popupWindowFeatures

AFFECTS PACKAGES:
@esri/arcgis-rest-auth